+++ module/Core/src/ShortUrl/Helper/ShortUrlTitleResolutionHelper.php
@@ -4,14 +4,19 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Core\ShortUrl\Helper;
 
+use Closure;
 use Fig\Http\Message\RequestMethodInterface;
 use GuzzleHttp\ClientInterface;
 use GuzzleHttp\RequestOptions;
+use Laminas\Stdlib\ErrorHandler;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Log\LoggerInterface;
 use Shlinkio\Shlink\Core\Config\Options\UrlShortenerOptions;
 use Throwable;
 
+use function function_exists;
 use function html_entity_decode;
+use function iconv;
 use function mb_convert_encoding;
 use function preg_match;
 use function str_contains;
@@ -30,9 +35,14 @@ readonly class ShortUrlTitleResolutionHelper implements ShortUrlTitleResolutionH
     // Matches the charset inside a Content-Type header
     private const string CHARSET_VALUE = '/charset=([^;]+)/i';
 
+    /**
+     * @param (Closure(): bool)|null $isIconvInstalled
+     */
     public function __construct(
         private ClientInterface $httpClient,
         private UrlShortenerOptions $options,
+        private LoggerInterface $logger,
+        private Closure|null $isIconvInstalled = null,
     ) {
     }
 
@@ -58,7 +68,7 @@ readonly class ShortUrlTitleResolutionHelper implements ShortUrlTitleResolutionH
         }
 
         $title = $this->tryToResolveTitle($response, $contentType);
-        return $title !== null ? $data->withResolvedTitle($title) : $data;
+        return $title !== null ? $data->withResolvedTitle(html_entity_decode(trim($title))) : $data;
     }
 
     private function fetchUrl(string $url): ResponseInterface|null
@@ -84,6 +94,7 @@ readonly class ShortUrlTitleResolutionHelper implements ShortUrlTitleResolutionH
     {
         $collectedBody = '';
         $body = $response->getBody();
+
         // With streaming enabled, we can walk the body until the </title> tag is found, and then stop
         while (! str_contains($collectedBody, '</title>') && ! $body->eof()) {
             $collectedBody .= $body->read(1024);
@@ -95,12 +106,48 @@ readonly class ShortUrlTitleResolutionHelper implements ShortUrlTitleResolutionH
             return null;
         }
 
-        // Get the page's charset from Content-Type header
-        preg_match(self::CHARSET_VALUE, $contentType, $charsetMatches);
+        $titleInOriginalEncoding = $titleMatches[1];
 
-        $title = isset($charsetMatches[1])
-            ? mb_convert_encoding($titleMatches[1], 'utf8', $charsetMatches[1])
-            : $titleMatches[1];
-        return html_entity_decode(trim($title));
+        // Get the page's charset from Content-Type header, or return title as is if not found
+        preg_match(self::CHARSET_VALUE, $contentType, $charsetMatches);
+        if (! isset($charsetMatches[1])) {
+            return $titleInOriginalEncoding;
+        }
+
+        $pageCharset = $charsetMatches[1];
+        return $this->encodeToUtf8WithMbString($titleInOriginalEncoding, $pageCharset)
+            ?? $this->encodeToUtf8WithIconv($titleInOriginalEncoding, $pageCharset)
+            ?? $titleInOriginalEncoding;
+    }
+
+    private function encodeToUtf8WithMbString(string $titleInOriginalEncoding, string $pageCharset): string|null
+    {
+        try {
+            return mb_convert_encoding($titleInOriginalEncoding, 'utf-8', $pageCharset);
+        } catch (Throwable $e) {
+            $this->logger->warning('It was impossible to encode page title in UTF-8 with mb_convert_encoding. {e}', [
+                'e' => $e,
+            ]);
+            return null;
+        }
+    }
+
+    private function encodeToUtf8WithIconv(string $titleInOriginalEncoding, string $pageCharset): string|null
+    {
+        $isIconvInstalled = ($this->isIconvInstalled ?? fn () => function_exists('iconv'))();
+        if (! $isIconvInstalled) {
+            $this->logger->warning('Missing iconv extension. Skipping title encoding');
+            return null;
+        }
+
+        try {
+            ErrorHandler::start();
+            $title = iconv($pageCharset, 'utf-8', $titleInOriginalEncoding);
+            ErrorHandler::stop(throw: true);
+            return $title ?: null;
+        } catch (Throwable $e) {
+            $this->logger->warning('It was impossible to encode page title in UTF-8 with iconv. {e}', ['e' => $e]);
+            return null;
+        }
     }
 }